+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Changelog
 * add missing params, update logic in login (thanks to xAranaktu #266)
 * reenable postion parsing & add missing keys in item_data parser (fix #265)
 * unify item_data keys for players & consumables
+* add playstyles & stadiums
 
 0.2.17 (2017-05-20)
 ^^^^^^^^^^^^^^^^^^^